refactor(Reorganize): Moved plugins into app-start. Combined view-matcher and utility.
