+++ server/app-start/hapi-sitemap/index.ts
