+++ server/app-start/hapi-sitemap/sitemap.xml.hbs
